Force adding modules that are explicitly part of the project to the cache.
Calling method.HasCustomDebugInfos eager loads the method body, which then (in some cases), eager loads dependant assemblies and adds them to the assembly cache. If one of those loaded assemblies is a project assembly, then the instance added to the cache will be one that dnlib auto-created. Meanwhile, the Confuser-created instance of that same assembly will be processed. This instance will not be equal to the one in the cache, which then results in the renaming service failing to add the correct references to be renamed. That then causes references to dependency to not be renamed in the parent assembly, resulting in an invalid binary.